second

+++ src/main/java/com/example/blog/services/serviceImpl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.blog.services.serviceImpl;
 
+import com.example.blog.dto.UsersDto;
 import com.example.blog.model.Friends;
 import com.example.blog.model.Users;
 import com.example.blog.repository.FriendsRepository;
@@ -11,12 +12,14 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
 @Service
 @Slf4j
@@ -36,16 +39,23 @@ public class UserServiceImpl implements UsersService {
     }
 
     public Page<Users> pageGetAllUsers(String byWhat, int next){
-        Pageable firstPageWithTwoElements = PageRequest.of(0, next, Sort.by(byWhat).ascending());
-        return userRepository.findAll(firstPageWithTwoElements);
+        Pageable firstPageElements = PageRequest.of(0, next, Sort.by(byWhat).ascending());
+        return userRepository.findAll(firstPageElements);
     }
 
-    public boolean addUser(Users user){
+    public boolean addUser(UsersDto user){
         Optional<Users> usersOptional = userRepository.findByEmail(user.getEmail());
         if(usersOptional.isPresent()){
             return true;
         }
-        userRepository.save(user);
+        Users user1 = new Users();
+        user1.setFirstName(user.getFirstName());
+        user1.setLastName(user.getLastName());
+        user1.setEmail(user.getEmail());
+        user1.setPhone(user.getPhone());
+        user1.setPassword(user.getPassword());
+        user1.setStatus(user.getStatus());
+        userRepository.save(user1);
         return false;
     }
 
@@ -57,6 +67,7 @@ public class UserServiceImpl implements UsersService {
         if (!user1.get().getPassword().equals(user.getPassword())) {
             return "Incorrect Password";
         }
+        user1.get().setStatus(true);
         return "valid";
     }
 
@@ -70,9 +81,18 @@ public class UserServiceImpl implements UsersService {
     }
 
 
-//    @Scheduled(fixedDelay = 10000)  // 10 Seconds
-    public void deleteUser(Long userId) {
+@Async
+    public CompletableFuture<String> deleteUser(Long userId) throws InterruptedException {
+        log.info("Counting down to delete");
+        Thread.sleep(30000L); // 30 sec
+        if (userRepository.findById(userId).get().getStatus()){
+            log.info("inside here");
+            log.info("executed");
+            return CompletableFuture.completedFuture("restored");
+        }
+        log.info("actually done deleting");
         userRepository.deleteById(userId);
+        return CompletableFuture.completedFuture("deleted");
     }
 
     @Override
